requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.1
django-crispy-forms==2.3
django-embed-video==1.4.10
django-sendgrid==1.0.1
idna==3.10
kombu==5.4.1
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.32.3
sendgrid==6.11.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
wheel==0.44.0
whitenoise==6.7.0
